chore: deprecate unusable databases
